+++ src/main/java/domain/participant/Player.java
@@ -1,0 +1,11 @@
+package domain.participant;
+
+import domain.Hand;
+
+
+public class Player extends Participant {
+
+    public Player(String name, Hand hand) {
+        super(name, hand);
+    }
+}
