+++ src/test/java/domain/constant/SuitTest.java
@@ -3,7 +3,6 @@ package domain.constant;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class SuitTest {
