+++ src/test/java/domain/participant/DealerTest.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.participant;
 
+import domain.Card;
 import org.junit.jupiter.api.Test;
 
 import static domain.constant.Rank.*;
